+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - [feature] New CLI command to generate migrations #6
+- [feature] Automatically sets database pool size based on listeners max_concurrency #2
 
 ## 0.1.1 (2016-09-02)
 - [bugfix] Handler is not injecting start and stop methods #4